requirements.txt
Django==5.1
python-decouple==3.8
pyTelegramBotAPI==4.23.0
mysql-connector-python==9.0.0
PyMySQL==1.1.1